refactor: seal class 적용
