+++ src/main/kotlin/blackjack/README.md
@@ -23,7 +23,7 @@
 - [x] Dealer가 둘이어서 어떤 딜러인지 헷갈릴 수도 있을 것 같은데요, 카드만 나눠주는 객체의 이름을 다르게 지어도 될 것 같습니다. 아니면 그냥 없애고 Deck을 활용할 수도 있을 것 같구요.
 - [x] 이 방식에는 버그가 있습니다. 10, A, A 의 카드를 받았다면 몇 점이어야할까요? soft 룰을 일반 언어로 표현하면 "A는 1점으로 계산하고, 손패에 A가 한 장이라도 있고 점수가 11 이하라면 점수에 +10을 한다"로 나타낼 수 있습니다. 이 방식의 구현이 로직적으로 맞더라도 이런 사실을 잘 드러내기 힘들어보여요. 복잡한 룰이 있다면 코드로 이 룰을 그대로 표현해보는 건 어떨까요?
 - [x] 요구사항을 잘 보시면 딜러의 첫 카드는 1장만 공개합니다. 이 것도 도메인 로직으로 표현해보시면 좋을 것 같아요.
-- [ ] 객체 내부에 타입을 enum으로 구분하는 경우라면 sealed class의 사용을 고려해볼 수 있을 것 같습니다.
+- [x] 객체 내부에 타입을 enum으로 구분하는 경우라면 sealed class의 사용을 고려해볼 수 있을 것 같습니다.
 - [ ] 이렇게 GameResult에서 결과를 계산하게 되면 테스트 코드가 살짝 애매해지는 것 같아요. 테스트를 구성할 때는 Player vs Dealer의 결과를 테스트 하는게 자연스러운 방법일 것 같은데 지금은 1대다가 아닌 1:1이더라도 Players를 만들어야하고, 결과도 Map을 받아서 테스트해야하니깐요. 테스트 코드도 클라이언트 코드이기 때문에 테스트가 자연스럽다면 객체가 자연스러운 책임을 갖게 될 수 있는 것 같습니다.
 - [ ] 블랙잭 룰을 읽어보시면 좋을 것 같네요. 플레이어의 순서는 딜러보다 항상 앞이기 때문에 플레이어가 버스트라면 항상 플레이어가 패배하게 됩니다.
 - [ ] 또 룰을 읽어보면 이를 bust라고 하는데요, getter를 사용하지 않고 Holder의 책임으로 나타낼 수 있을 것 같습니다. 도메인을 탐구하고 이해하게 되면 코드도 훨씬 깔끔해질 수 있는 것 같습니다.
